Quickspace: Add Quick events support
This adds a new UI to Quickspace when events are available such as Now
Playing, Alarms, or whatever we configure as an event. This logic is
inspired by Google's event infrastructure in Smartspace but completely
written from scratch to be housed in Quickspace.

Initial events are:

- Device intro: Shows on first boot and setup
- Ambient Play (Now Playing): Shows what song is playing and tapping
the song will launch the results in Google search

spezi77 edits:
- Fix package name for DoubleShadowTextView
- Update strings for ROM intro
- Update weather client package

Bootleggers edit:
- Let's disable the Now Playing code, might recycle it later

Signed-off-by: spezi77 <spezi7713@gmx.net>
[kenny3fcb: Remove unused code]
Signed-off-by: Subhajeet Muhuri <kenny3fcb@gmail.com>
Change-Id: I2742c13043b816c54bc2cefdfd415f52f2e76d3f
Signed-off-by: Sebastian <kjjjnob.seba@gmail.com>